Translate expressions to R(base) and R(data.table)
Generalise the tidyverse emitter into one R emitter parameterised by
dialect, adding the R(base) and R(data.table) targets the spec lists.
All three share R's core semantics (three-valued logic, %%, the %in%
null guard, date promotion, NaN notes); the dialects differ only in
spelling:

- Strings: base/data.table use nchar, tolower/toupper, trimws,
  startsWith/endsWith, and grepl(..., perl = TRUE) with a PCRE note.
  Only grepl needs a null guard; startsWith/endsWith propagate NA.
- Selections: only the tidyverse keeps if_all(); base and data.table
  expand to a conjunction, per the spec.
- Aggregates: data.table gets .N and uniqueN(x, na.rm = TRUE); base
  gets length(unique(x[!is.na(x)])) and refuses ROW_COUNT(), which a
  bare predicate in base R cannot name.
- CASE/conditionals: case_when / nested ifelse / fcase; kind tests
  use if_else / ifelse / fifelse.
- BETWEEN: base expands to x >= lo & x <= hi; data.table uses its
  native between().

--target r now resolves to R(base), the spec's family default. The
spec's example output is aligned with the emitters' integer literals
(10L). The R emitter tests live in r.rs, organised by construct with
the dialect variants side by side where they differ.
